pyproject.toml
[tool.poetry]
name = "social_network"
version = "0.1.0"
description = ""
authors = ["Vladislav Sytiuk <vladislav.sytiuk@chisw.com>"]

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "3.4.1"
certifi = "2021.10.8"
cffi = "1.15.0"
charset-normalizer = "2.0.10"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "36.0.1"
defusedxml = "0.7.1"
Django = "4.0.1"
django-filter = "21.1"
django-templated-mail = "1.1.1"
djangorestframework = "3.13.1"
djangorestframework-simplejwt = "4.8.0"
Faker = "11.3.0"
idna = "3.3"
itypes = "1.2.0"
Jinja2 = "3.0.3"
MarkupSafe = "2.0.1"
oauthlib = "3.1.1"
pycparser = "2.21"
PyJWT = "2.3.0"
python-dateutil = "2.8.2"
python3-openid = "3.2.0"
pytz = "2021.3"
requests = "2.27.1"
requests-oauthlib = "1.3.0"
six = "1.16.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.1.0"
sqlparse = "0.4.2"
text-unidecode = "1.3"
tzdata = "2021.5"
uritemplate = "4.1.1"
urllib3 = "1.26.8"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
black = "^22.8.0"
pytest = "^7.1.3"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
